Remade the whole shizzle
Instead of just using nmap (which is cringe) I wanted to try and make it as bare bones as possible, which meant trying to make a scanning algorithm myself.
I made it using a for loop of pings and for the port scans I used netcat. Now this method sucks ass and is incredibly slow... but it works, somewhat.